+++ src/hal/hal_SysTick.h
@@ -8,7 +8,7 @@ void hal_SysTick_intRegister(void (*const fun)(void));
 void hal_SysTick_intUnregister(void);
 void hal_SysTick_intEnable(void);
 void hal_SysTick_intDisable(void);
-void hal_SysTick_setPeriodMillis(uint32_t const periodMillis);
-uint32_t hal_SysTick_getPeriodMillis(void);
+void hal_SysTick_setPeriodMillis(uint16_t const periodMillis);
+uint16_t hal_SysTick_getPeriodMillis(void);
 
 #endif // HAL_SYSTICK_H
